replace `simd` feature with `no_simd`
`simd` does not make a useful feature, because there is no way for a
user to be dependent on it being enabled (I don't expose any SIMD-only
interfaces). Rather, `no_simd` is a feature, because some users can't
compile SIMD intrinsics for various reasons. Fixes #25.

This has the additional benefit of eliminating a default-feature, a
facility I have found to be a Cargo misfeature.