Derive zerocopy traits via macro, remove custom derive dependency
This commit disables zerocopy's `derive` feature, replacing uses of
custom derives with uses of a newly-introducedi [1] `macro_rules!` macro
which is specifically added to support the ppv-lite86 and groestl-aesni
crates. This allows us to no longer take a dependency on any proc macro
derive, and thus to no longer take dependencies on expensive-to-compile
crates like syn.

[1] https://github.com/google/zerocopy/pull/2418